requirements.txt
openai
langchain
streamlit
pypdf
pinecone-client
sentence_transformers